Revert "refactor: Three.js 파일 디렉터리 구조 수정"
This reverts commit 92d57ea3e3a9cc8363469329390fe80410ceef47.
